+++ server/src/main/java/com/ninebudget/service/InstitutionService.java
@@ -6,13 +6,16 @@ import com.ninebudget.model.mapper.InstitutionMapper;
 import com.ninebudget.repository.InstitutionRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.LinkedList;
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 /**
  * Service Implementation for managing {@link Institution}.
@@ -48,14 +51,17 @@ public class InstitutionService {
     /**
      * Get all the institutions.
      *
-     * @param pageable the pagination information.
      * @return the list of entities.
      */
     @Transactional(readOnly = true)
-    public Page<InstitutionDto> findAll(Pageable pageable) {
+    public List<InstitutionDto> findAll() {
         log.debug("Request to get all Institutions");
-        return institutionRepository.findAll(pageable)
-            .map(institutionMapper::toDto);
+        User principal = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        return institutionRepository.findAll().stream()
+                .filter(account -> account.getAccount().getId().toString().equals(principal.getUsername()))
+                .map(institutionMapper::toDto)
+                .collect(Collectors.toCollection(LinkedList::new));
     }
 
 
@@ -68,8 +74,11 @@ public class InstitutionService {
     @Transactional(readOnly = true)
     public Optional<InstitutionDto> findOne(UUID id) {
         log.debug("Request to get Institution : {}", id);
+        User principal = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
         return institutionRepository.findById(id)
-            .map(institutionMapper::toDto);
+                .filter(account -> account.getAccount().getId().toString().equals(principal.getUsername()))
+                .map(institutionMapper::toDto);
     }
 
     /**
@@ -79,6 +88,12 @@ public class InstitutionService {
      */
     public void delete(UUID id) {
         log.debug("Request to delete Institution : {}", id);
-        institutionRepository.deleteById(id);
+        User principal = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        //Only delete if User has access
+        Optional<Institution> trans = institutionRepository.findById(id);
+        if (trans.isPresent() && trans.get().getAccount().getId().toString().equals(principal.getUsername())) {
+            institutionRepository.deleteById(id);
+        }
     }
 }
