+++ server/src/main/java/com/ninebudget/service/TransactionService.java
@@ -6,13 +6,16 @@ import com.ninebudget.model.mapper.TransactionMapper;
 import com.ninebudget.repository.TransactionRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.LinkedList;
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 /**
  * Service Implementation for managing {@link Transaction}.
@@ -48,14 +51,17 @@ public class TransactionService {
     /**
      * Get all the transactions.
      *
-     * @param pageable the pagination information.
      * @return the list of entities.
      */
     @Transactional(readOnly = true)
-    public Page<TransactionDto> findAll(Pageable pageable) {
+    public List<TransactionDto> findAll() {
         log.debug("Request to get all Transactions");
-        return transactionRepository.findAll(pageable)
-            .map(transactionMapper::toDto);
+        User principal = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        return transactionRepository.findAll().stream()
+                .filter(trans -> trans.getAccount().getId().toString().equals(principal.getUsername()))
+                .map(transactionMapper::toDto)
+                .collect(Collectors.toCollection(LinkedList::new));
     }
 
 
@@ -68,8 +74,11 @@ public class TransactionService {
     @Transactional(readOnly = true)
     public Optional<TransactionDto> findOne(UUID id) {
         log.debug("Request to get Transaction : {}", id);
+        User principal = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
         return transactionRepository.findById(id)
-            .map(transactionMapper::toDto);
+                .filter(trans -> trans.getAccount().getId().toString().equals(principal.getUsername()))
+                .map(transactionMapper::toDto);
     }
 
     /**
@@ -79,6 +88,12 @@ public class TransactionService {
      */
     public void delete(UUID id) {
         log.debug("Request to delete Transaction : {}", id);
-        transactionRepository.deleteById(id);
+        User principal = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        //Only delete if User has access
+        Optional<Transaction> trans = transactionRepository.findById(id);
+        if (trans.isPresent() && trans.get().getAccount().getId().toString().equals(principal.getUsername())) {
+            transactionRepository.deleteById(id);
+        }
     }
 }
